+++ CHANGELOG.md
@@ -175,6 +175,11 @@ TODO
   - Add local ISD/AS to Path
 - Renamed "List" methods in PathMetadata API
   [#223](https://github.com/scionproto-contrib/jpan/pull/223)
+- Endhost API improvements
+  [#224](https://github.com/scionproto-contrib/jpan/pull/224)
+  TODO 
+  - fix fixed ISD/AS for DNSUtil in DatagramSocketApiTest 
+  
 
 ### TODO
   - Test! -> Receive paths with different ISD?!?! -> separate PR?